requirements.txt
pandas
numpy
pathlib
seaborn
matplotlib
ta
pytrends
numpy
scipy
praw
python-dotenv